Implement delegation of tasks to agents. Hardcoded test. Added procedure describtion

+++ src/main/java/Components/PlanProposal.java
@@ -6,18 +6,19 @@ public class PlanProposal extends Message {
     ArrayList<Action> actions;
     Agent a;
     long taskID;
-    int startIndex;
-    int endIndex;
+    long startIndex;
+    long endIndex;
 
-    public PlanProposal(ArrayList<Action> actions, Agent a, long taskID, int startIndex, int endIndex) {
-        this.actions = actions;
-        this.a = a;
-        this.taskID = taskID;
-        this.startIndex = startIndex;
-        this.endIndex = endIndex;
+    public PlanProposal(ArrayList<Action> actions, Agent a, long taskID, long startIndex, long endIndex) {
+	this.actions = actions;
+	this.a = a;
+	this.taskID = taskID;
+	this.startIndex = startIndex;
+	this.endIndex = endIndex;
     }
 
     public String toString() {
-        return "Agent " + a.getAgentNumber() + " has plan for task  " + taskID + " with start index " + startIndex + " and end index " + endIndex + " and the actions: " + actions.toString();
+	return "Agent " + a.getAgentNumber() + " has plan for task  " + taskID + " with start index " + startIndex
+		+ " and end index " + endIndex + " and the actions: " + actions.toString();
     }
 }
